--- a/pyproject.toml
+++ b/pyproject.toml
@@ -46,6 +46,7 @@ ulid = "^1.1"
 sqlalchemy-bigquery = { version = "1.9.0", python = "<3.13" }
 airbyte-api = "^0.49.2"
 google-cloud-bigquery-storage = "^2.25.0"
+uuid7 = "^0.1.0"
 
 [tool.poetry.group.dev.dependencies]
 docker = "^7.0.0"