--- a/pyproject.toml
+++ b/pyproject.toml
@@ -40,7 +40,6 @@ snowflake-connector-python = "^3.10.0"
 snowflake-sqlalchemy = "^1.5.1"
 sqlalchemy = "1.4.51"
 types-pyyaml = "^6.0.12.12"
-ulid = "^1.1"
 
 # TODO: Remove this arbitrary python constraint once `sqlalchemy-bigquery` has done so.
 sqlalchemy-bigquery = { version = "1.9.0", python = "<3.13" }