## Summary

- Improved unique ID generation by switching from ULID to a custom UUID method, ensuring more robust and unique identifier creation.
Chores

- Updated dependencies in pyproject.toml, including the addition of the uuid7 and types-pyyaml packages, while removing ulid.

In the previous flamegraph, the ULID creation was taking approx. 4us, plus up to 5us to convert to `str` type. The `UUID v7` implementation generates a UUID in one step in approx 2us. This cuts off approximately 7us out of ~26us processing time per record, or up to 30% of the per-record processing time.

Now (2us of 19us total):

<img width="436" alt="image" src="https://github.com/user-attachments/assets/ce8101d8-b5cc-427a-849e-a89e26bbcd56">

Previous (4+5us, combined 9us of 27us total):

<img width="377" alt="image" src="https://github.com/user-attachments/assets/8cc4f873-e155-4cfc-9346-5f8b8ebeb931">



Notes:

- ULIDs are still used for batch IDs, file name suffixes, and table name suffixes. This change only affects unique record IDs.
- In working this PR, I found that there was an unused dependency on `ulid` library, where we are/were actually using `python-ulid`.
- A nice added benefit is that the row IDs are now compatible with the UUID column type.

New benchmark (41K records / sec using the `e2e` source):

`poetry run python ./examples/run_perf_test_reads.py -e=6 --source=e2e`

<img width="641" alt="image" src="https://github.com/user-attachments/assets/e32a4e24-2a45-4e5c-9e62-a52a6530c0c4">